Added firebase auth support for signup view
